requirements.txt
streamlit
numpy
pandas
matplotlib
seaborn
scikit-learn
regex
wordcloud
underthesea
imblearn
nltk
demoji
pyvi
mysql-connector-python==8.2.0
